+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v2.1.0
+* Add support for automatically reloading the extension when the "Apama Home" configuration option is changed.
 
 ## v2.0.0